+++ metasrc/tex/scraper.py
@@ -1,0 +1,101 @@
+"""TeX source scraping.
+"""
+
+import re
+
+
+# Regular expressions for "\def \name {content}"
+# Expects the entire command to be on one line.
+DEF_PATTERN = re.compile(
+    r"\\def\s*"  # def command with optional whitespace
+    r"(?P<name>\\[a-zA-Z]*?)\s*"  # macro name with optional whitespace
+    r"{(?P<content>.*?)}")  # macro contents
+
+# Regular expressions for "\def \name {content}"
+# Does not handle arguments.
+NEWCOMMAND_PATTERN = re.compile(
+    r"\\newcommand\s*"  # def command with optional whitespace
+    r"{\s*(?P<name>\\[a-zA-Z]*?)\s*}\s*"  # macro name with optional whitespace
+    r"{(?P<content>.*?)}")  # macro contents
+
+
+def get_macros(tex_source):
+    """Get all macro definitions from TeX source, supporting multiple
+    declaration patterns.
+
+    Parameters
+    ----------
+    tex_source : `str`
+        TeX source content.
+
+    Returns
+    -------
+    macros : `dict`
+        Keys are macro names (including leading ``\``) and values are the
+        content (as `str`) of the macros.
+
+    Notes
+    -----
+    This function uses the following function to scrape macros of different
+    types:
+
+    - `get_def_macros`
+    - `get_newcommand_macros`
+
+    This macro scraping has the following caveats:
+
+    - Macro definition (including content) must all occur on one line.
+    - Macros with arguments are not supported.
+    """
+    macros = {}
+    macros.update(get_def_macros(tex_source))
+    macros.update(get_newcommand_macros(tex_source))
+    return macros
+
+
+def get_def_macros(tex_source):
+    """Get all ``\def`` macro definition from TeX source.
+
+    Parameters
+    ----------
+    tex_source : `str`
+        TeX source content.
+
+    Returns
+    -------
+    macros : `dict`
+        Keys are macro names (including leading ``\``) and values are the
+        content (as `str`) of the macros.
+
+    Notes
+    -----
+    ``\def`` macros with arguments are not supported.
+    """
+    macros = {}
+    for match in DEF_PATTERN.finditer(tex_source):
+        macros[match.group('name')] = match.group('content')
+    return macros
+
+
+def get_newcommand_macros(tex_source):
+    r"""Get all ``\newcommand`` macro definition from TeX source.
+
+    Parameters
+    ----------
+    tex_source : `str`
+        TeX source content.
+
+    Returns
+    -------
+    macros : `dict`
+        Keys are macro names (including leading ``\``) and values are the
+        content (as `str`) of the macros.
+
+    Notes
+    -----
+    ``\newcommand`` macros with arguments are not supported.
+    """
+    macros = {}
+    for match in NEWCOMMAND_PATTERN.finditer(tex_source):
+        macros[match.group('name')] = match.group('content')
+    return macros